chore: split controllers into multiple files
