+++ app/Http/Controllers/TakeClaimController.php
@@ -7,7 +7,7 @@ use App\Exceptions\ClaimNotFoundException;
 use App\Exceptions\ClaimTakenException;
 use Illuminate\Contracts\View\View;
 
-class ClaimController extends Controller
+class TakeClaimController extends Controller
 {
     public function checkClaim(string $claimId): View
     {
@@ -26,25 +26,6 @@ class ClaimController extends Controller
         return view('claim')
             ->with('claimId', $claim->getKey())
             ->with('copy', $claim->getSuccessfulText());
-    }
-
-    public function createClaimForm(): View
-    {
-        return view('create-claim');
-    }
-
-    public function createClaim(): View
-    {
-        $successfulText = request()->get('successful_claim_text');
-        $unsuccessfulText = request()->get('unsuccessful_claim_text');
-
-        $newClaim = Claim::create([
-            'successful_claim_text' => $successfulText,
-            'unsuccessful_claim_text' => $unsuccessfulText,
-        ]);
-
-        return view('create-claim-success')
-            ->with('claimId', $newClaim->getKey());
     }
 
     protected function getClaim(string $claimId): Claim
